third_party/blink/web_tests/platform/generic/fast/dom/Range/surroundContents-for-detached-node-expected.txt

PASS surroundContents(sample.cloneNode(true)) is "<b>2345</b>"
PASS surroundContents(sample) is "<b>2345</b>"
PASS successfullyParsed is true

TEST COMPLETE
